+++ Unigram/Unigram/Services/ContactsService.cs
@@ -36,7 +36,7 @@ namespace Unigram.Services
         private readonly DisposableMutex _syncLock;
         private readonly object _importedPhonesRoot;
 
-        private int[] _contacts;
+        private long[] _contacts;
 
         private CancellationTokenSource _syncToken;
 
@@ -50,7 +50,7 @@ namespace Unigram.Services
             _syncLock = new DisposableMutex();
             _importedPhonesRoot = new object();
 
-            _contacts = new int[0];
+            _contacts = new long[0];
 
             _aggregator.Subscribe(this);
         }
@@ -258,7 +258,7 @@ namespace Unigram.Services
                 return;
             }
 
-            var remove = new List<int>();
+            var remove = new List<long>();
 
             var existing = _contacts;
             if (existing != null)
@@ -452,17 +452,17 @@ namespace Unigram.Services
             }
         }
 
-        public static Task<int?> GetContactIdAsync(Contact contact)
+        public static Task<long?> GetContactIdAsync(Contact contact)
         {
             if (contact == null)
             {
-                return Task.FromResult<int?>(null);
+                return Task.FromResult<long?>(null);
             }
 
             return GetContactIdAsync(contact.Id);
         }
 
-        public static async Task<int?> GetContactIdAsync(string contactId)
+        public static async Task<long?> GetContactIdAsync(string contactId)
         {
             var annotationStore = await ContactManager.RequestAnnotationStoreAsync(ContactAnnotationStoreAccessType.AppAnnotationsReadWrite);
             var store = await ContactManager.RequestStoreAsync(ContactStoreAccessType.AppContactsReadWrite);
@@ -485,7 +485,7 @@ namespace Unigram.Services
                     }
 
                     var remote = first.RemoteId;
-                    if (int.TryParse(remote.Substring(1), out int userId))
+                    if (long.TryParse(remote.Substring(1), out long userId))
                     {
                         return userId;
                     }
